Well I made a mess and should have made more commits and an additional branch to work in during all this updated work.
I made a mess I need to clean up, compact, refactor and automate.

Lots of new code can be put into functions or class methods.

Some class variables are no longer needed.

the use of references and deleting has to be redone as it doens't work
much.


tons of changes:
new memory manager that adds cells of memory of everthing that is
allocated with new and freed with delete.
other changes around the code but mostly spagahetti as I tried fixing a
fuew bugs that led me down a tangent. But after lots of additions and
reversals I have an existing codebase from which to continue from.

Test harnesses /unit tests still need to be created for all classes,
methods, functions, etc.
Readjustments to classes and cleanup soon to follow.

+++ src/Environment.c
@@ -8,22 +8,31 @@
 #include "CException.h"
 
 #include <stdio.h>
+#include <assert.h>
 #include "Environment.h"
 #include "hash.h"
 #include "HashMap.h"
 #include "Token.h"
 char* func(void* value){
 	if(value){
-	    char* num ;
+	    char* num,*num2 ;
 	    Object * val = (Object*)value;
 	    if(val->type == NUMBER){
 		   num = NULL;
 		   asprintf(&num,"%f",val->value.number);
-		   return num;
+		   num2 = (char*)new(RAW,sizeof(char)*(strlen(num)+1));
+		   memset(num2,0,strlen(num)+1);
+		   strncpy(num2,num,strlen(num));
+		   free(num);
+		   num = NULL;
+		   return num2;
 	    }
 	    else{
 		   num = NULL;
-		   return strdup(val->value.string);
+		   num =new(RAW,sizeof(char)*(strlen(val->value.string)+1));
+		   memset(num,0,strlen(val->value.string)+1);
+		   strncpy(num,val->value.string,strlen(val->value.string));
+		   return num;
 	    }
 	}
 	return NULL;
@@ -37,9 +46,12 @@ void* get(Environment* env,Token* name){
 	char* temp;
     temp = NULL;
 	map = env->hashMap;
-	get_node = env->hashMap->super.vtable.get_node_for_key;
+/*    if(strcmp(name->lexeme,"fibonacci")==0){
+	   fprintf(stderr,"found in get\n");
+    }*/
+	get_node = map->super.vtable.get_node_for_key;
 	if((node = (HashMapNode*)get_node((struct _HASH*)map,name->lexeme))){
-		return node->super.value;
+		return getReference(node->super.value);
 	}
     if(env->Enclosing != NULL){
 	   void * temp;
@@ -47,17 +59,23 @@ void* get(Environment* env,Token* name){
 	   return temp;
     }
 	temp = NULL;
+    temp = new(RAW,sizeof(char)*(strlen("Undefined variable ''.")+strlen(name->lexeme)+1));
+    memset(temp,0,(strlen("Undefined variable ''.")+strlen(name->lexeme))+1);
 	asprintf(&temp,"Undefined variable '%s'.",name->lexeme);
     e.id = 9;
-	e.token = name;
+	e.token = (name);
 	e.message = temp;
+    e.sub = NULL;
 	Throw(e);
     return NULL;
 /*	Throw(runtimeError(name,);*/
 }
 
 void defineEnv(Environment* env,char* name, Object *value){
-	env->hashMap->super.vtable.add_to_hash((struct _HASH*)env->hashMap,strdup(name),value);
+/*    if(strcmp(name,"fibonacci")==0){
+	   fprintf(stderr,"found in defineEnv\n");
+    }*/
+	env->hashMap->super.vtable.add_to_hash((struct _HASH*)env->hashMap,name,value);
 }
 
 void init_Environment(Environment* env){
@@ -65,12 +83,24 @@ void init_Environment(Environment* env){
 	env->defineEnv= &defineEnv;
 	env->assign = &assign;
     env->get = &get;
+    env->copy = &copy_Environment;
     env->Enclosing = NULL;
 }
 void deleteEnvironment(Environment* env){
-    delete_hashmap((struct _HASH*)env->hashMap);
-    env->hashMap = NULL;
-    env->Enclosing = NULL;
+    if(getReferenceCount(env) <= 1){
+	   mem_footer* footer;
+	   footer = get_footer(env);
+	   delete_hashmap((struct _HASH*)env->hashMap);
+	   env->hashMap = NULL;
+	   env->Enclosing = NULL;
+	   footer = get_footer(env);
+	   if(footer->functions.allocated){
+		  delete(env);
+	   }
+    }
+    else {
+	   releaseReference(env);
+    }
 }
 
 void init_EnvironmentwithEnclosing(Environment* env,Environment* enclosing){
@@ -92,10 +122,23 @@ void assign(Environment* env, Token* name, Object* value){
 	   return;
     }
 	temp = NULL;
+	temp = new(RAW,sizeof(char)*(strlen("Undefined variable ''.")+strlen(name->lexeme)+1));
 	asprintf(&temp,"Undefined variable '%s'.",name->lexeme);
     e.id = 10;
-	e.token = name;
+	e.token = getReference(name);
 	e.message = temp;
+	e.sub = NULL;
 	Throw(e);
     return;
+}
+
+void* copy_Environment(void* inE){
+    Environment* out, *in;
+    in = (Environment*)inE;
+    if(!in) return NULL;
+    out = new(OBJECTIVE,sizeof(Environment));
+    init_Environment(out);
+    out->Enclosing = getReference(in->Enclosing);
+    out->hashMap = copy(in->hashMap);
+    return out;
 }
